requirements.txt
Django==5.2.3
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
celery==5.3.6
redis==5.0.1
selenium==4.16.0
webdriver-manager==4.0.1
Pillow==10.1.0
psutil==5.9.6
gunicorn==21.2.0
whitenoise==6.6.0
django-extensions==3.2.3
